Wrong Screen for Nanox plus Minor fixes
- NanoX and NanoS Svgs were flipped
- Wrong stroke width on NanoX
- Duplicated icon removed and refactored the old one to change sizes
- Cleanup
