+++ src/screens/Onboarding/steps/get-started.js
@@ -1,6 +1,6 @@
 // @flow
 import React, { Component, PureComponent } from "react";
-import { StyleSheet, View, Linking } from "react-native";
+import { StyleSheet, View } from "react-native";
 import { Trans } from "react-i18next";
 import Icon from "react-native-vector-icons/dist/Feather";
 
@@ -10,12 +10,9 @@ import OnboardingLayout from "../OnboardingLayout";
 import LText from "../../../components/LText";
 import Touchable from "../../../components/Touchable";
 import { withOnboardingContext } from "../onboardingContext";
-import IconImport from "../../../icons/Import";
 import IconCheck from "../../../icons/Check";
 import IconRestore from "../../../icons/History";
 import colors from "../../../colors";
-import { urls } from "../../../config/urls";
-import { deviceNames } from "../../../wording";
 
 const IconPlus = () => <Icon name="plus" color={colors.live} size={16} />;
 
@@ -44,13 +41,6 @@ class OnboardingStepGetStarted extends Component<OnboardingStepProps> {
     return (
       <OnboardingLayout header="OnboardingStepGetStarted">
         <TrackScreen category="Onboarding" name="GetStarted" />
-
-        <LText style={styles.subtitle} secondary semiBold>
-          <Trans
-            i18nKey="onboarding.stepGetStarted.withNanoX"
-            values={deviceNames.nanoX}
-          />
-        </LText>
         <Row
           id="initialize"
           Icon={IconPlus}
@@ -69,32 +59,6 @@ class OnboardingStepGetStarted extends Component<OnboardingStepProps> {
           label={<Trans i18nKey="onboarding.stepGetStarted.initialized" />}
           onPress={this.onInitialized}
         />
-
-        <LText style={[styles.subtitle, styles.extraMargin]} secondary semiBold>
-          <Trans
-            i18nKey="onboarding.stepGetStarted.withoutNanoX"
-            values={deviceNames.nanoX}
-          />
-        </LText>
-        <Row
-          id="import"
-          Icon={IconImport}
-          label={<Trans i18nKey="onboarding.stepGetStarted.import" />}
-          onPress={this.onImport}
-        />
-        <Touchable
-          event="BuyNanoX"
-          style={styles.footer}
-          onPress={() => Linking.openURL(urls.buyNanoX)}
-        >
-          <LText style={styles.footerText} semiBold>
-            <Trans
-              i18nKey="onboarding.stepGetStarted.buy"
-              values={deviceNames.nanoX}
-            />
-          </LText>
-          <Icon size={16} name="chevron-right" color={colors.live} />
-        </Touchable>
       </OnboardingLayout>
     );
   }
